requirements.txt
streamlit
pillow
pdfrw
requests
openai
python-dotenv
